Rename constants folder to errors folder
